feat(zql): Add support for take after join
