+++ packages/zql/src/ivm/take.ts
@@ -238,16 +238,15 @@ export class Take implements Operator {
     }
 
     assert(this.#output, 'Output not set');
-    // When take below join is supported, this assert should be removed
-    // and a 'child' change should be pushed to output if its row
-    // is <= bound.
-    assert(change.type !== 'child', 'child changes are not supported');
 
     const {takeState, takeStateKey, maxBound, constraint} =
-      this.#getStateAndConstraint(change.node.row);
+      this.#getStateAndConstraint(
+        change.type === 'child' ? change.row : change.node.row,
+      );
     if (!takeState) {
       return;
     }
+
     const {compareRows} = this.getSchema();
 
     if (change.type === 'add') {
@@ -379,6 +378,12 @@ export class Take implements Operator {
         maxBound,
       );
       this.#output.push(change);
+    } else if (change.type === 'child') {
+      // A 'child' change should be pushed to output if its row
+      // is <= bound.
+      if (takeState.bound && compareRows(change.row, takeState.bound) <= 0) {
+        this.#output.push(change);
+      }
     }
   }
